rename solver modules
