+++ chirho/dynamical/handlers/trajectory.py
@@ -4,7 +4,7 @@ import pyro
 import torch
 
 from chirho.dynamical.internals._utils import append
-from chirho.dynamical.internals.backend import simulate_trajectory
+from chirho.dynamical.internals.solver import simulate_trajectory
 from chirho.dynamical.ops import Trajectory
 
 T = TypeVar("T")
