--- chirho/dynamical/internals/solver/__init__.py
@@ -1,1 +1,0 @@
-from . import torchdiffeq  # noqa: F401
